+++ engine/Cargo.toml
@@ -13,3 +13,4 @@ tokio-tungstenite = "0.24"
 serde = { version = "1.0 ", features = ["derive"] }
 serde_json = "1.0"
 futures = "0.3"
+lazy_static = "1.5.0"
